update public_meetings_cal.rb
Update from Surender Dalal. The update parses and creates new events for RECURRING events (which the old script did not). The old script used to only pull the FIRST event of a recurring events. Now, for recurring events with no end date, we create copies as new events for up to 12 months into the future from the Day that the job runs based on the recurrence logic of the original event.